+++ resources/views/user-column.blade.php
@@ -1,47 +1,37 @@
 @php
-use Filament\Support\Icons\Heroicon;
-
 $state = $getState();
 $size = $getSize();
+$stackedModalAction = $getStackedModalAction();
+$isStackedModalAction = $stackedModalAction && ($getAction() === $stackedModalAction);
 @endphp
 
-<div  class="fi-user-entry fi-size-{{ $size }}"
-      {{-- x-data
-      x-load-css="[@js(\Filament\Support\Facades\FilamentAsset::getStyleHref('filament-user-field', package: 'deldius/filament-user-field'))]" --}}
-      >
-  @if ($state)
-    @if ($getShowAvatar())
-      <div style="position: relative">
-        @if ($avatarUrl = $getAvatarUrl())
-          <img class="fi-user-entry-avatar fi-size-{{ $size }}" src="{{ $avatarUrl }}" alt="User Avatar">
-        @else
-          <div class="fi-user-entry-default-avatar fi-size-{{ $size }}">
-            <x-filament::icon :icon="Heroicon::User" class="fi-size-{{ $size }}"/>
-          </div>
-        @endif
+<div class="fi-user-entry fi-size-{{ $size }}">
+  @if ($isStackedState() && $state->isNotEmpty())
+    <div class="fi-user-stack fi-size-{{ $size }}" aria-label="Users">
+      @if ($isStackedModalAction)
+        <span class="fi-sr-only">Open user list for {{ $getLabel() ?: 'Users' }}</span>
+      @endif
 
-        @if ($getShowActiveState())
-          @if ($getIsActiveState())
-            <div class="fi-user-entry-active-state fi-size-{{ $size }}" style="color: green">
-              <x-filament::icon :icon="Heroicon::OutlinedCheckCircle"/>
-            </div>
-          @else
-            <div class="fi-user-entry-active-state fi-size-{{ $size }}" style="color: red">
-              <x-filament::icon :icon="Heroicon::OutlinedXCircle"/>
-            </div>
-          @endif
-        @endif
-      </div>
-    @endif
+      @foreach ($getVisibleStackedUsers() as $user)
+        <div class="fi-user-stack-item">
+          @include('filament-user-field::components.user-avatar', [
+              'avatarUrl' => $getAvatarUrlFor($user),
+              'size' => $size,
+              'alt' => $user->{config('user-field.user_model.fields.heading', 'name')} ?? 'User',
+          ])
+        </div>
+      @endforeach
 
-    <div class="fi-user-entry-content">
-      <div class="fi-user-entry-content-heading fi-size-{{ $size }}">
-        <span class="">{{ $getHeading() }}</span>
-      </div>
-      <div class="fi-user-entry-content-description fi-size-{{ $size }}">{{ $getDescription() }}</div>
+      @if ($remaining = $getStackedRemainingCount())
+        <span class="fi-user-stack-remaining fi-size-{{ $size }}">+{{ $remaining }}</span>
+      @endif
     </div>
+  @elseif ($state && ! $isStackedState())
+    @include('filament-user-field::components.user-card', [
+        'state' => $state,
+        'size' => $size,
+    ])
   @else
-  {{-- Empty State --}}
     <div>
       @if ($emptyState = $getEmptyState())
         {{ $emptyState }}