+++ src/Concerns/HasState.php
@@ -3,7 +3,9 @@
 namespace Deldius\UserField\Concerns;
 
 use App\Models\User;
+use DateInterval;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Cache;
 
 trait HasState
@@ -14,6 +16,18 @@ trait HasState
     {
         $state = parent::getState();
 
+        if (is_array($state) || $state instanceof Collection) {
+            return collect($state)
+                ->map(fn (mixed $item): mixed => $this->resolveUserState($item))
+                ->filter(fn (mixed $item): bool => $item !== null)
+                ->values();
+        }
+
+        return $this->resolveUserState($state);
+    }
+
+    protected function resolveUserState(mixed $state): mixed
+    {
         /** @disregard P1009 */
         // @phpstan-ignore class.notFound
         $userModel = config('user-field.user_model.class', User::class);
@@ -23,22 +37,16 @@ trait HasState
             return $state;
         }
 
-        if ($state) {
-            $data = Cache::remember(
-                key: self::CACHE_KEY_PREFIX . $userModel . '_' . $state,
-                ttl: new \DateInterval('PT5S'), // 5 seconds
-                callback: function () use ($userModel, $userModelId, $state) {
-                    return $userModel::where($userModelId, $state)->first()?->toArray();
-                }
-            );
-
-            if ($data === null) {
-                return null;
-            }
-
-            return new $userModel($data);
+        if ((! is_scalar($state)) || (! $state)) {
+            return null;
         }
 
-        return null;
+        $data = Cache::remember(
+            key: self::CACHE_KEY_PREFIX . $userModel . '_' . $state,
+            ttl: new DateInterval('PT5S'),
+            callback: fn () => $userModel::where($userModelId, $state)->first()?->toArray(),
+        );
+
+        return ($data === null) ? null : new $userModel($data);
     }
 }