+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `filament-user-field` will be documented in this file.
 
+## Unreleased
+
+- Add automatic avatar stacks for array and Collection user state.
+- Add configurable stack limits and opt-in user-list modals to `UserEntry` and `UserColumn`.
+
 ## 1.0.0 - 202X-XX-XX
 
 - initial release